pyproject.toml
[project]
name = "github-repo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-anthropic>=1.4.1",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
]
